Creating the file
This windows form application has a text box input and will display the entered value on the form window. This also shows the current system date and there is button which will open another form window and there is a button which can rotate the image shown inside that popup window.